Added python function open()
Returns the proper Dataset type corresponding to the given dtype,
or uses the dtype in the attributes.json file if dtype is not provided.
Note there is now one more test that fails.
It is possible to create two Datasets of different types using the same
file system. Writing multple datatypes can cause problems reading from
the dataset.