Make tests pass.
Failures are expected since the behavior of N5 when writting multiple
datatypes is not quite clear, and failure cases for floats have not
been properly explored yet.

python 3.5 version of shutils does not accept a pathlib.Path object,
so I extract it into a string.

tox now passes 3.5 and 3.6 locally